fix invalid formatspecifier.
The problem was that printing an uint64_t with the functions of the printf-family correctly is hard.
This commit adds an explicit static cast to unsigned long long to remove both the resulting UB and the warning.

+++ src/cpp/generators/test/GeneratorsGTest.cpp
@@ -176,7 +176,7 @@ TEST_F(GeneratorsGTest, testDynamicPubWebGenerator) {
 
 		// output in clustered way for better visual inspection
 		char path[23];
-		sprintf(path, "output/pubweb-%04llu.eps", i);
+		sprintf(path, "output/pubweb-%04llu.eps", static_cast<unsigned long long>(i));
 		TRACE("path: " , path);
 		PostscriptWriter psWriter(G, true);
 		psWriter.write(path);
